Rewrite a lot of logic to handle a missing Content-Range header
This change rewrites the fetch logic to handle the case where the `Content-Range`
header is not exposed. When this is the case, the full resource needs to be used.
Fetch responses are now wrapped in a `Usable<T>` or `Unusable<T>` type to allow
bubbling the response back to the top-level function.

A note about the deleted test: it is not useful to test the case where the server
does not support range requests AND doesn't expose the `Content-Range` header, as
that doesn't matter—if range requests are not supported we don't and shouldn't
bother with any other logic, just bubble the full resource to the top.